+++ R/helpfunctions.R
@@ -167,7 +167,7 @@ if(getRversion() >= "2.15.1") {
           "u.mean", "v.mean", "write.csv", "x", "y", "z", ".", "time2",
           "group", "step", "point", "change", "end", "level", "m", "rotate",
           "x.d", "y.d", "PC", "step2", "runif", "N", "angle", "var", "head",
-          "col__", "row__", "t1", "z1", "z2"))
+          "col__", "row__", "t1", "z1", "z2", "..nn"))
 }